Move choice list logic into Choices Handle sorting the choice list into columns inside the Choices component.
